fix 64bit compilation problems

+++ trunk/engine/src/persistance/TFactory.h
@@ -36,7 +36,7 @@ public:
    TFactoryList ();
    ~TFactoryList ();
 
-   typedef int FactoryID;
+   typedef size_t FactoryID;
 
    bool add (TFactoryBase*);
    bool remove (const char*);
